# flags2: use pointer receivers to avoid copying structs all around