Inspired by your PR, it seems like we can improve perf here by using pointer receivers (and explicit reference creation in loops):

```
name                     old time/op    new time/op    delta
GoforitEnabledOn           12.5ns ± 2%    12.2ns ± 1%   -2.75%  (p=0.008 n=5+5)
GoforitEnabledOn-2         34.6ns ± 5%    33.7ns ± 3%     ~     (p=0.421 n=5+5)
GoforitEnabledOn-4         42.3ns ±18%    37.2ns ±12%     ~     (p=0.151 n=5+5)
GoforitEnabledOn-8         61.3ns ± 1%    58.3ns ± 5%   -5.01%  (p=0.008 n=5+5)
GoforitEnabledOff          11.4ns ± 0%    11.5ns ± 0%   +1.21%  (p=0.016 n=4+5)
GoforitEnabledOff-2        38.6ns ±10%    37.0ns ± 6%     ~     (p=0.421 n=5+5)
GoforitEnabledOff-4        46.3ns ±12%    51.9ns ±18%     ~     (p=0.690 n=5+5)
GoforitEnabledOff-8        66.5ns ± 2%    66.2ns ± 0%     ~     (p=0.556 n=5+4)
GoforitEnabledPartial      74.8ns ± 1%    63.6ns ± 0%  -14.97%  (p=0.008 n=5+5)
GoforitEnabledPartial-2    55.0ns ± 5%    48.1ns ± 2%  -12.42%  (p=0.008 n=5+5)
GoforitEnabledPartial-4    52.4ns ±15%    45.1ns ± 3%  -13.89%  (p=0.008 n=5+5)
GoforitEnabledPartial-8    68.2ns ± 2%    68.7ns ± 3%     ~     (p=0.556 n=5+4)
```

The benchmark is in a different repo, and includes testing the wrapper we use.  All tests in this repo, as well as the benchmark pass with the race detector on.